fixed: error under Android
On Android autoref count is used. So, any implicit conversion of class
type to TObject class leads to AV.
So, on Android we cannot store class type as TObject reference.

To avoid this, let's introduce classes to store explicit records.

Also, mobile compilers have 0-based strings indexing, so use
ZEROBASEDSTRINGS conditional define to make the code to work on Android.